+++ packages/atomjs/src/app.cjs
@@ -7,6 +7,7 @@ const fs = require('node:fs');
 const state = require('./state.cjs');
 const { BridgeServer } = require('./bridge-server.cjs');
 const { stopNativeHost } = require('./native-host.cjs');
+const { stopWindowsNativeHost } = require('./windows-native-host.cjs');
 
 class App extends EventEmitter {
   constructor() {
@@ -43,6 +44,7 @@ class App extends EventEmitter {
 
     for (const win of [...state.windows.values()]) win.destroy();
     await stopNativeHost();
+    await stopWindowsNativeHost();
     if (state.bridgeServer) await state.bridgeServer.stop();
     this.emit('will-quit');
     this.emit('quit', {}, 0);